Calculate air lead exposure estimates at specific locations.

calculate_airPb() uses a land use random forest model developed by Dr. Cole Brokamp based on ambient air sampling in Cincinnati, OH between 2001 and 2005 to estimate exposure to airborne lead at point locations in the area specified by latitude and longitude. The model predictors include greenspace (NDVI) within 1000 meters, population density within 500 meters, length of bus routes within 900 meters, percent pasure within 800 meters, percent developed open land within 1100 meters, percent developed medium land within 400 meters, percent developed low land within 900 meters, and percent developed high land within 1500 meters.

calculate_airPb(locations, return.LU.vars = FALSE)

Arguments

locations
Data.frame with columns 'id', 'lat', and 'lon' at minimum.

return.LU.vars
When return.LU.vars = TRUE, the land use predictors used to generate the air lead values are also returned.

Value

If return.LU.vars = FALSE, a numeric vector of air lead estimates (ug/m3) ? is returned. If return.LU.vars = TRUE, the locations data.frame with additional columns for air lead values and the land use predictors used to generate the air lead values is returned.

References

Cole Brokamp, Roman Jandarov, MB Rao, Grace LeMasters, Patrick Ryan. Exposure assessment models for elemental components of particulate matter in an urban environment: A comparison of regression and random forest approaches. Atmospheric Environment. 151. 1-11. 2017. http://dx.doi.org/10.1016/j.atmosenv.2016.11.066

Examples

my_data <- data.frame(id = 1:3,
    lat = c(39.19674, 39.12731, 39.28765),
    lon = c(-84.58260, -84.52700, -84.51017))

lead_est <- calculate_airPb(my_data, return.LU.vars = FALSE)
lead_est <- calculate_airPb(my_data, return.LU.vars = TRUE)